move value types to a separate namespace
