+++ config/packages/doctrine_mongodb.yaml
@@ -18,11 +18,11 @@ doctrine_mongodb:
           alias: Siklid
   types:
     email:
-      class: App\Foundation\ValueObject\EmailType
+      class: App\Foundation\ValueType\EmailType
     slug:
-      class: App\Foundation\ValueObject\SlugType
+      class: App\Foundation\ValueType\SlugType
     username:
-      class: App\Foundation\ValueObject\UsernameType
+      class: App\Foundation\ValueType\UsernameType
 
 when@prod:
   doctrine_mongodb:
